move 20.qna-with-appinsights, fix eslint errors
