--- javascript_nodejs/22.qna-with-appinsights/deploymentScripts/azuredeploy.json
@@ -1,1 +1,0 @@
-{}
