+++ samples/javascript_nodejs/20.qna-with-appinsights/.env
@@ -1,0 +1,2 @@
+botFilePath=./qna-with-appinsights.bot
+botFileSecret=
