+++ samples/javascript_nodejs/20.qna-with-appinsights/cognitiveModels/smartLightFAQ.tsv
