+++ samples/javascript_nodejs/20.qna-with-appinsights/deploymentScripts/msbotClone/bot.recipe
@@ -1,0 +1,38 @@
+{
+  "version": "1.0",
+  "resources": [
+    {
+      "type": "endpoint",
+      "id": "1",
+      "name": "development",
+      "url": "http://localhost:3978/api/messages"
+    },
+    {
+      "type": "luis",
+      "id": "34",
+      "name": "nlp-with-luis-LUIS"
+    },
+    {
+      "type": "endpoint",
+      "id": "2",
+      "name": "production",
+      "url": "https://your-bot-url.azurewebsites.net/api/messages"
+    },
+    {
+      "type": "abs",
+      "id": "3",
+      "name": "nlp-with-luis-abs"
+    },
+    {
+      "type": "appInsights",
+      "id": "4",
+      "name": "nlp-with-luis-insights"
+    },
+    {
+      "type": "blob",
+      "id": "5",
+      "name": "nlp-with-luis-blob",
+      "container": "botstatestore"
+    }
+  ]
+}
