+++ samples/javascript_nodejs/20.qna-with-appinsights/index.js
@@ -1,46 +1,41 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-const { BotFrameworkAdapter } = require("botbuilder");
-const { BotConfiguration } = require("botframework-config");
-const path = require("path");
-const restify = require("restify");
-const { QnAMakerBot } = require("./bot");
-const { MyAppInsightsMiddleware } = require("./middleware");
-
-const CONFIG_ERROR = 1;
-
-// Bot name as defined in .bot file.
-// See https://aka.ms/about-bot-file to learn more about .bot files.
-const BOT_CONFIGURATION = 'qna-with-appinsights';
-const QNA_CONFIGURATION = 'qnamakerService';
-const APP_INSIGHTS_CONFIGURATION = 'appInsights';
+const path = require('path');
+const restify = require('restify');
+const { BotFrameworkAdapter } = require('botbuilder');
+const { BotConfiguration } = require('botframework-config');
+const { QnAMakerBot } = require('./bot');
+const { MyAppInsightsMiddleware } = require('./middleware');
 
 // Read botFilePath and botFileSecret from .env file.
 // Note: Ensure you have a .env file and include botFilePath and botFileSecret.
 const ENV_FILE = path.join(__dirname, './.env');
 const env = require('dotenv').config({ path: ENV_FILE });
 
-// Create HTTP server.
-let server = restify.createServer();
-server.listen(process.env.port || process.env.PORT || 3978, function () {
-    console.log(`\n${server.name} listening to ${server.url}.`);
-    console.log(`\nGet Bot Framework Emulator: https://aka.ms/botframework-emulator.`);
-    console.log(`\nTo talk to your bot, open qnamaker-with-app-insights.bot file in the emulator.`);
-});
-
-// .bot file path.
+// Get the .bot file path.
 const BOT_FILE = path.join(__dirname, (process.env.botFilePath || ''));
-
 let botConfig;
 try {
-    // Read bot configuration from .bot file. 
+    // Read bot configuration from .bot file.
     botConfig = BotConfiguration.loadSync(BOT_FILE, process.env.botFileSecret);
+} catch (err) {
+    console.error(`\nError reading bot file. Please ensure you have valid botFilePath and botFileSecret set for your environment.`);
+    console.error(`\n - The botFileSecret is available under appsettings for your Azure Bot Service bot.`);
+    console.error(`\n - If you are running this bot locally, consider adding a .env file with botFilePath and botFileSecret.\n\n`);
+    process.exit();
 }
-catch (err) {
-    console.log(`Error reading bot file. Please ensure you have valid botFilePath and botFileSecret set for your environment.`);
-    process.exit(CONFIG_ERROR);
-}
+
+// For local development configuration as defined in .bot file.
+const DEV_ENVIRONMENT = 'development';
+
+// Bot name as defined in .bot file or from runtime.
+// See https://aka.ms/about-bot-file to learn more about .bot files.
+const BOT_CONFIGURATION = (process.env.NODE_ENV || DEV_ENVIRONMENT);
+
+// QnA Maker and Application Insights service names as found in .bot file.
+const QNA_CONFIGURATION = 'qnamakerService';
+const APP_INSIGHTS_CONFIGURATION = 'appInsights';
 
 // Get bot endpoint configuration by service name.
 const endpointConfig = botConfig.findServiceByNameOrId(BOT_CONFIGURATION);
@@ -79,12 +74,32 @@ const adapter = new BotFrameworkAdapter({
 // This cached TelemetryClient is used by the custom class MyAppInsightsQnAMaker.
 adapter.use(new MyAppInsightsMiddleware(appInsightsSettings));
 
-// Create the main dialog.
-const qnaMakerBot = new QnAMakerBot(qnaEndpointSettings, {}, logMessage, logName);
+// Catch-all for errors.
+adapter.onTurnError = async (turnContext, error) => {
+    console.error(`\n [onTurnError]: ${ error }`);
+    await turnContext.sendActivity(`Oops. Something went wrong!`);
+};
+
+// Create the SuggestedActionsBot.
+let bot;
+try {
+    bot = new QnAMakerBot(qnaEndpointSettings, {}, logMessage, logName);
+} catch (err) {
+    console.error(`[botInitializationError]: ${ err }`);
+    process.exit();
+}
+
+// Create HTTP server.
+let server = restify.createServer();
+server.listen(process.env.port || process.env.PORT || 3978, function() {
+    console.log(`\n${ server.name } listening to ${ server.url }.`);
+    console.log(`\nGet Bot Framework Emulator: https://aka.ms/botframework-emulator.`);
+    console.log(`\nTo talk to your bot, open qnamaker-with-app-insights.bot file in the emulator.`);
+});
 
 // Listen for incoming requests.
 server.post('/api/messages', async (req, res) => {
     await adapter.processActivity(req, res, async (context) => {
-        await qnaMakerBot.onTurn(context);
+        await bot.onTurn(context);
     });
 });
