+++ samples/javascript_nodejs/20.qna-with-appinsights/middleware/index.js
@@ -3,4 +3,4 @@
 
 const { MyAppInsightsMiddleware } = require('./myAppInsightsMiddleware');
 
-exports.MyAppInsightsMiddleware = MyAppInsightsMiddleware;
+module.exports.MyAppInsightsMiddleware = MyAppInsightsMiddleware;
