+++ samples/javascript_nodejs/20.qna-with-appinsights/package.json
@@ -12,13 +12,19 @@
   "license": "MIT",
   "dependencies": {
     "applicationinsights": "^1.0.4",
-    "botbuilder": "^4.0.0-preview1.1135",
-    "botbuilder-ai": "^4.0.0-preview1.1135",
-    "botframework-config": "^4.0.0-preview1.1102",
+    "botbuilder": "^4.0.6",
+    "botbuilder-ai": "^4.0.6",
+    "botframework-config": "^4.0.6",
     "dotenv": "^6.0.0",
     "restify": "^6.4.0"
   },
   "devDependencies": {
+    "eslint": "^5.6.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-node": "^7.0.1",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-standard": "^4.0.0",
     "nodemon": "^1.18.4"
   }
 }
