+++ samples/javascript_nodejs/20.qna-with-appinsights/qna-with-appinsights.bot
